requirements.txt
# Core dependencies
PySide6==6.6.1
python-socketio==5.11.0
aiohttp==3.9.1
python-dotenv==1.0.1
psutil==5.9.8
fire==0.5.0
pywinauto==0.6.8

# Development dependencies
pytest>=7.3.1
black>=22.12.0
flake8>=6.0.0
mypy>=1.0.0

# Additional dependencies
alembic==1.13.1
asyncio==3.4.3
markdown==3.5.2
qasync==0.27.1
PyYAML==6.0.1
socketio==0.2.1
SQLAlchemy==2.0.25
Lugwit_Module==1.0.0  # 请确保这个模块可用
